Features: modules and new initiaization
* Now we have two new modules: utilities.py contains all the function
  not   directly related to the optimization problem, optimization.py
  instead contains the key function for the resolution of the problem
* Created function printPath to print a path in a file as a list of the
  nodes visited, printed in visit order

Bugfix: the problem was badly formulated, without the constraint on the
number of vehicles the optimal solution is using all the dummy path. So,
we add some constraints to ESP subproblem in order to make him generate
long feasible paths, then we use that paths to initialize oure master
problem. That is a bit redundant and there is for sure a smarter way to
do that (a smart way to initialize dual variables without solving the
master problem, for example), but for now we go for the brutal way.

Author: Simone Richetti <richetti.s@gmail.com>