Bugfix: ESP now works
Modifying the reduced costs formula and the way we compute M the ESP
problem now works. Now, we know we cannot use it, because it takes
an infinity of time (actually I don't know precisely how much) to solve
one single iteration of the problem. We have to implement the dynamic
programming algorithm.

Author: Simone Richetti <richetti.s@gmail.com>